changed stock qty and price to decimal ipv double

+++ Libraries/DigiBugzy.Data/Migrations/InitialCreate.cs
@@ -346,13 +346,13 @@ namespace DigiBugzy.Data.Migrations
             creatory.StartNewTable(_currentTableName);
             creatory.CreateBaseAdministrationEntity();
             creatory.AddColumn(nameof(StockJournal.EntryDate), isNullable: false, fieldType: BaseEntityCreator.FieldTypes.AsDateTime);
-            creatory.AddColumn(nameof(StockJournal.QuantityIn), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.QuantityOut), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.QuantityOnOrder), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.TotalInStock), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.QuantityReserved), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.Price), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
-            creatory.AddColumn(nameof(StockJournal.TotalValue), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDouble);
+            creatory.AddColumn(nameof(StockJournal.QuantityIn), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.QuantityOut), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.QuantityOnOrder), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.TotalInStock), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.QuantityReserved), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.Price), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
+            creatory.AddColumn(nameof(StockJournal.TotalValue), isNullable: true, fieldType: BaseEntityCreator.FieldTypes.AsDecimal);
 
             //ProjectSectionPartId
             creatory.AddColumn(nameof(StockJournal.ProjectSectionPartId), isNullable: true);
